+++ tinyms/model/bert/bert.py
@@ -27,66 +27,6 @@ from tinyms import Tensor
 from tinyms import Parameter
 
 
-class BertConfig:
-    """
-    Configuration for `BertModel`.
-
-    Args:
-        seq_length (int): Length of input sequence. Default: 128.
-        vocab_size (int): The shape of each embedding vector. Default: 32000.
-        hidden_size (int): Size of the bert encoder layers. Default: 768.
-        num_hidden_layers (int): Number of hidden layers in the BertTransformer encoder
-                           cell. Default: 12.
-        num_attention_heads (int): Number of attention heads in the BertTransformer
-                             encoder cell. Default: 12.
-        intermediate_size (int): Size of intermediate layer in the BertTransformer
-                           encoder cell. Default: 3072.
-        hidden_act (str): Activation function used in the BertTransformer encoder
-                    cell. Default: "gelu".
-        hidden_dropout_prob (float): The dropout probability for BertOutput. Default: 0.1.
-        attention_probs_dropout_prob (float): The dropout probability for
-                                      BertAttention. Default: 0.1.
-        max_position_embeddings (int): Maximum length of sequences used in this
-                                 model. Default: 512.
-        type_vocab_size (int): Size of token type vocab. Default: 16.
-        initializer_range (float): Initialization value of TruncatedNormal. Default: 0.02.
-        use_relative_positions (bool): Specifies whether to use relative positions. Default: False.
-        dtype (:class:`mindspore.dtype`): Data type of the input. Default: mstype.float32.
-        compute_type (:class:`mindspore.dtype`): Compute type in BertTransformer. Default: mstype.float32.
-    """
-    def __init__(self,
-                 seq_length=128,
-                 vocab_size=32000,
-                 hidden_size=768,
-                 num_hidden_layers=12,
-                 num_attention_heads=12,
-                 intermediate_size=3072,
-                 hidden_act="gelu",
-                 hidden_dropout_prob=0.1,
-                 attention_probs_dropout_prob=0.1,
-                 max_position_embeddings=512,
-                 type_vocab_size=16,
-                 initializer_range=0.02,
-                 use_relative_positions=False,
-                 dtype=mstype.float32,
-                 compute_type=mstype.float32):
-        self.seq_length = seq_length
-        self.vocab_size = vocab_size
-        self.hidden_size = hidden_size
-        self.num_hidden_layers = num_hidden_layers
-        self.num_attention_heads = num_attention_heads
-        self.hidden_act = hidden_act
-        self.intermediate_size = intermediate_size
-        self.hidden_dropout_prob = hidden_dropout_prob
-        self.attention_probs_dropout_prob = attention_probs_dropout_prob
-        self.max_position_embeddings = max_position_embeddings
-        self.type_vocab_size = type_vocab_size
-        self.initializer_range = initializer_range
-        self.use_relative_positions = use_relative_positions
-        self.dtype = dtype
-        self.compute_type = compute_type
-
-
 class EmbeddingLookup(layers.Layer):
     """
     A embeddings lookup table with a fixed dictionary and size.
